refactor: useVoteSocket와 추상화 수준 통일
- 소켓 연결/해제를 제외한 소켓 이벤트는 useVoteSocketEvents에서 처리
